services/replay-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
redis==5.0.1
